Improved control over remote volume mounts using mount policies
Mount policies, that affects how the telepresence traffic-agent shares
the pod's volumes, and also how the client will mount them, can now be
provided using the Helm chart value `agent.mountPolicies` or as JSON
object in the workload annotation `telepresence.io/mount-policies`.
A mount policy is applied to a volume or to all paths matching a
path-prefix (distinguished by checking if first character is a '/'), and
can be one of `Ignore`, `Local`, `Remote`, or `RemoteReadOnly`.

Signed-off-by: Thomas Hallgren <thomas@tada.se>