Make telepresence connect --docker work with Rancher Desktop
Rancher Desktop will start a K3s control-plane and typically expose the
Kubernetes API server at `127.0.0.1:6443`. Telepresence can connect to
this cluster when running on the host, but the address is not available
when connecting in docker mode.

The problem is solved by ensuring that the Kubernetes API server address
used when doing a `telepresence connect --docker` is swapped from
127.0.0.1 to the internal address of the control-plane node. This works
because that address is available to other docker containers, and the
Kubernetes API server is configured with a certificate that accepts it.

Signed-off-by: Thomas Hallgren <thomas@tada.se>