Add a telepresence wiretap command.
A wiretap is a read-only form of an intercept. The original container
will still receive the wiretapped traffic, and a copy of the traffic
is sent to the client. Writes on the client connection are discarded.

Similar to an ingest, a wiretap will always enforce read-only status on
all volume mounts, and since the wiretap then is completely read-only,
there's no limit to how many simultaneous wiretaps that can be served.

Signed-off-by: Thomas Hallgren <thomas@tada.se>